3.2.1
- Added predicted spawning.
- Added PredictedSpawn component.
- Added several features to PredictionManager.
- Changed moved several prediction datas and callbacks from TimeManager to PredictionManager.
- Changed ObjectId internally from short to ushort.
- Changed PrefabId internally from short to ushort.
- Added TImeManager.LocalTickToTick().
- Added Reader.Clear().
- Added NetworkConnection.UNSET_CLIENTID_VALUE.
- Added TransportManager.GetLowestMTU(byte).
- Added NetworkObject.PredictedSpawner.
- Added NetworkObject.PredictedSpawn.
- Added NetworkObject.UNSET_OBJECTID_VALUE.
- Added NetworkObject.UNSET_PREFABID_VALUE.
- Improved DefaultObjectPool and ObjectPool now supports addressables.
- Improved PredictedObject smoothing.
- Fixed PredictedObject sending Rigidody2D velocity as a Vector3 instead of Vector2.
- Fixed incorrect prefab collections being used with addressables.
- Fixed timed observer conditions not working on nested NetworkObjects.
- Fixed incorrect Reader/Writer being called when serializing multiple nullable types.
- Fixed SyncVar OnChange callbacks ignoring overrides.
- Fixed conn.Objects missing objects when calling GiveOwnership to the same connection that is already owner.
- Fixed incorrect packing reference in NetworkTransform.
- Improved XML in several areas.
- Improved NetworkBehaviours now log when automatically adding the NetworkObject component.
- Fixed Visual Studio code cleanup removing needed logic.
- Fixed typo in QuaternionFN class name.

+++ Assets/FishNet/Runtime/Managing/Object/PrefabObjects/DualPrefabObjects.cs
@@ -110,8 +110,8 @@ namespace FishNet.Managing.Object
         {
             for (int i = startIndex; i < _prefabs.Count; i++)
             {
-                ManagedObjects.InitializePrefab(_prefabs[i].Server, (short)i, CollectionId);
-                ManagedObjects.InitializePrefab(_prefabs[i].Client, (short)i, CollectionId);
+                ManagedObjects.InitializePrefab(_prefabs[i].Server, i, CollectionId);
+                ManagedObjects.InitializePrefab(_prefabs[i].Client, i, CollectionId);
             }
         }
