+++ Assets/FishNet/Runtime/Managing/Object/PrefabObjects/SinglePrefabObjects.cs
@@ -104,7 +104,7 @@ namespace FishNet.Managing.Object
         public override void InitializePrefabRange(int startIndex)
         {
             for (int i = startIndex; i < _prefabs.Count; i++)
-                ManagedObjects.InitializePrefab(_prefabs[i], (short)i, CollectionId);
+                ManagedObjects.InitializePrefab(_prefabs[i], i, CollectionId);
         }
